pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yamlr"
version = "0.1.0"
description = "Yamlr: High-fidelity Kubernetes manifest healing engine."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Nishar A Sunkesala", email = "maintainers@yamlr.dev"}
]
keywords = ["kubernetes", "yaml", "devops", "sre", "yamlr", "manifest-healing"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

# External Dependencies
dependencies = [
    "ruamel.yaml>=0.17.21",
    "rich>=12.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "mypy>=1.0",
]

# For creating the standalone binaries for Yamlr
build = ["pyinstaller>=6.4.0"]

# Standard Entry Point
[project.scripts]
yamlr = "yamlr.cli.main:main"

[project.urls]
Homepage = "https://yamlr.dev"
Issues = "https://github.com/yamlr-dev/yamlr/issues"
Repository = "https://github.com/yamlr-dev/yamlr.git"

# --- SETUPTOOLS CONFIGURATION ---
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
# Explicitly exclude pro modules from OSS builds
# This ensures "src/yamlr/pro" is not packaged in the OSS wheel
exclude = ["tests*", "catalog*", "yamlr.pro*"]
namespaces = true

[tool.setuptools.package-data]
# Include the K8s catalogs in the distribution
yamlr = ["catalog/*.json"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]


